refactor: separate data classes to evosuite and llm folders
